HV: Seed module restructure
As part of arch module cleanup, moving out seed
module as it has no dependency on arch.

Acked-by: Jason Chen <jason.cj.chen@intel.com>
