--- hypervisor/arch/x86/seed/seed_abl.h
@@ -1,12 +1,0 @@
-/*
- * Copyright (C) 2018 Intel Corporation. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef SEED_ABL_H_
-#define SEED_ABL_H_
-
-bool parse_seed_abl(uint64_t addr, struct physical_seed *phy_seed);
-
-#endif /* SEED_ABL_H_ */
