+++ hypervisor/crypto/seed.c
@@ -14,8 +14,6 @@
 #include <multiboot.h>
 #include <crypto_api.h>
 #include <seed.h>
-#include "seed_abl.h"
-#include "seed_sbl.h"
 
 #define BOOTLOADER_SBL  0U
 #define BOOTLOADER_ABL  1U
@@ -27,6 +25,13 @@ struct seed_argument {
 	uint64_t addr;
 };
 
+/* Structure of physical seed */
+struct physical_seed {
+	struct seed_info seed_list[BOOTLOADER_SEED_MAX_ENTRIES];
+	uint32_t num_seeds;
+	uint32_t pad;
+};
+
 #define SEED_ARG_NUM 4U
 static struct seed_argument seed_arg[SEED_ARG_NUM] = {
 	{ "ImageBootParamsAddr=",        BOOTLOADER_SBL, 0UL },
@@ -36,6 +41,200 @@ static struct seed_argument seed_arg[SEED_ARG_NUM] = {
 };
 
 static struct physical_seed g_phy_seed;
+
+#define SEED_ENTRY_TYPE_SVNSEED         0x1U
+/* #define SEED_ENTRY_TYPE_RPMBSEED        0x2U */
+
+/* #define SEED_ENTRY_USAGE_USEED          0x1U */
+#define SEED_ENTRY_USAGE_DSEED          0x2U
+
+struct seed_list_hob {
+	uint8_t revision;
+	uint8_t reserved0[3];
+	uint32_t buffer_size;
+	uint8_t total_seed_count;
+	uint8_t reserved1[3];
+};
+
+struct seed_entry {
+	/* SVN based seed or RPMB seed or attestation key_box */
+	uint8_t type;
+	/* For SVN seed: useed or dseed
+	 * For RPMB seed: serial number based or not
+	 */
+	uint8_t usage;
+	/* index for the same type and usage seed */
+	uint8_t index;
+	uint8_t reserved;
+	/* reserved for future use */
+	uint16_t flags;
+	/* Total size of this seed entry */
+	uint16_t seed_entry_size;
+	/* SVN seed: struct seed_info
+	 * RPMB seed: uint8_t rpmb_key[key_len]
+	 */
+	uint8_t seed[0];
+};
+
+struct image_boot_params {
+	uint32_t size_of_this_struct;
+	uint32_t version;
+	uint64_t p_seed_list;
+	uint64_t p_platform_info;
+	uint64_t reserved;
+};
+
+#define ABL_SEED_LEN 32U
+struct abl_seed_info {
+	uint8_t svn;
+	uint8_t reserved[3];
+	uint8_t seed[ABL_SEED_LEN];
+};
+
+#define ABL_SEED_LIST_MAX 4U
+struct abl_svn_seed {
+	uint32_t size_of_this_struct;
+	uint32_t version;
+	uint32_t num_seeds;
+	struct abl_seed_info seed_list[ABL_SEED_LIST_MAX];
+};
+
+/*
+ * parse_seed_abl
+ *
+ * description:
+ *    This function parse seed_list which provided by ABL.
+ *
+ * input:
+ *    cmdline       pointer to cmdline string
+ *
+ * output:
+ *    phy_seed      pointer to physical seed structure
+ *
+ * return value:
+ *    true if parse successfully, otherwise false.
+ */
+static bool parse_seed_abl(uint64_t addr, struct physical_seed *phy_seed)
+{
+	uint32_t i;
+	uint32_t legacy_seed_index = 0U;
+	struct seed_info *seed_list;
+	struct abl_svn_seed *abl_seed = (struct abl_svn_seed *)hpa2hva(addr);
+	bool status = false;
+
+	if ((phy_seed != NULL) && (abl_seed != NULL) &&
+	    (abl_seed->num_seeds >= 2U) && (abl_seed->num_seeds <= ABL_SEED_LIST_MAX)) {
+
+		seed_list = phy_seed->seed_list;
+		/*
+		 * The seed_list from ABL contains several seeds which based on SVN
+		 * and one legacy seed which is not based on SVN. The legacy seed's
+		 * svn value is minimum in the seed list. And CSE ensures at least two
+		 * seeds will be generated which will contain the legacy seed.
+		 * Here find the legacy seed index first.
+		 */
+		for (i = 1U; i < abl_seed->num_seeds; i++) {
+			if (abl_seed->seed_list[i].svn < abl_seed->seed_list[legacy_seed_index].svn) {
+				legacy_seed_index = i;
+			}
+		}
+
+		/*
+		 * Copy out abl_seed for trusty and clear the original seed in memory.
+		 * The SOS requires the legacy seed to derive RPMB key. So skip the
+		 * legacy seed when clear original seed.
+		 */
+		(void)memset((void *)&phy_seed->seed_list[0U], 0U, sizeof(phy_seed->seed_list));
+		for (i = 0U; i < abl_seed->num_seeds; i++) {
+			seed_list[i].cse_svn = abl_seed->seed_list[i].svn;
+			(void)memcpy_s((void *)&seed_list[i].seed[0U], sizeof(seed_list[i].seed),
+					(void *)&abl_seed->seed_list[i].seed[0U], sizeof(abl_seed->seed_list[i].seed));
+
+			if (i == legacy_seed_index) {
+				continue;
+			}
+
+			(void)memset((void *)&abl_seed->seed_list[i].seed[0U], 0U,
+							sizeof(abl_seed->seed_list[i].seed));
+		}
+
+		phy_seed->num_seeds = abl_seed->num_seeds;
+		status = true;
+	}
+
+	return status;
+}
+
+/*
+ * parse_seed_sbl
+ *
+ * description:
+ *    This function parse seed_list which provided by SBL
+ *
+ * input:
+ *    cmdline       pointer to cmdline string
+ *
+ * return value:
+ *    true if parse successfully, otherwise false.
+ */
+static bool parse_seed_sbl(uint64_t addr, struct physical_seed *phy_seed)
+{
+	uint8_t i;
+	uint8_t dseed_index = 0U;
+	struct image_boot_params *boot_params = NULL;
+	struct seed_list_hob *seed_hob = NULL;
+	struct seed_entry *entry = NULL;
+	struct seed_info *seed_list = NULL;
+	bool status = false;
+
+	boot_params = (struct image_boot_params *)hpa2hva(addr);
+
+	if (boot_params != NULL) {
+		seed_hob = (struct seed_list_hob *)hpa2hva(boot_params->p_seed_list);
+	}
+
+	if ((seed_hob != NULL) && (phy_seed != NULL)) {
+		status = true;
+
+		seed_list = phy_seed->seed_list;
+
+		entry = (struct seed_entry *)((uint8_t *)seed_hob + sizeof(struct seed_list_hob));
+
+		for (i = 0U; i < seed_hob->total_seed_count; i++) {
+			if (entry != NULL) {
+				/* retrieve dseed */
+				if ((SEED_ENTRY_TYPE_SVNSEED == entry->type) &&
+				    (SEED_ENTRY_USAGE_DSEED == entry->usage)) {
+
+					/* The seed_entry with same type/usage are always
+					 * arranged by index in order of 0~3.
+					 */
+					if ((entry->index != dseed_index) ||
+					    (entry->index >= BOOTLOADER_SEED_MAX_ENTRIES)) {
+						status = false;
+						break;
+					}
+
+					(void)memcpy_s((void *)&seed_list[dseed_index], sizeof(struct seed_info),
+							(void *)&entry->seed[0U], sizeof(struct seed_info));
+					dseed_index++;
+
+					/* erase original seed in seed entry */
+					(void)memset((void *)&entry->seed[0U], 0U, sizeof(struct seed_info));
+				}
+
+				entry = (struct seed_entry *)((uint8_t *)entry + entry->seed_entry_size);
+			}
+		}
+
+		if (status) {
+			phy_seed->num_seeds = dseed_index;
+		}
+	}
+
+	return status;
+}
+
 
 static uint32_t parse_seed_arg(void)
 {
